Moves subtitlesettinglabel to it's own file
